chore: Move envConfig to a constant folder #3
